requirements.txt
streamlit-nightly
pandas
requests
joblib
pydeck
scikit-learn
plotly
numpy